refactor(components): Move Auth components
- Move auth components in a shared folder
- Clean custom components body
- Delete unused Navbar and PageWrapper
