+++ src/shared/components/Auth/common/Logo.tsx
