+++ src/shared/components/Auth/common/index.ts
