+++ src/shared/components/Auth/index.ts
@@ -1,0 +1,5 @@
+import CustomConfirmSignUp from "./ConfirmSignUp";
+import CustomSignIn from "./SignIn";
+import CustomSignUp from "./SignUp";
+
+export { CustomConfirmSignUp, CustomSignIn, CustomSignUp };
